+++ CHANGELOG.md
@@ -132,6 +132,11 @@
 - When compiling to JavaScript constants will now be annotated as `@__PURE__`.
   ([Giacomo Cavalieri](https://github.com/giacomocavalieri))
 
+- The `little` and `big` endianness options, the `signed` and `unsigned` integer
+  options, and sized floats (32-bit and 64-bit), can now be used in bit array
+  expressions and patterns on the JavaScript target.
+  ([Richard Viney](https://github.com/richard-viney))
+
 ### Formatter
 
 ### Language Server